Add function to equalize left/right region statistics
A new function `equalizeLeftRightRegions` was added to the `label_analysis.py` file. The function ensures that corresponding left and right regions have the same mean and standard deviation values in the training data. This should help when creating training data as we don't want the random distributions to be different for corresponding left and right regions.